pyproject.toml
[tool.poetry]
name = "ollama-rename-img"
version = "0.1.0"
description = ""
authors = ["Tedfulk <teddyfulk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
pydantic = "^2.8.2"
ollama = "^0.3.0"
typer = "^0.12.3"
tqdm = "^4.66.4"

[tool.poetry.scripts]
rename = "ollama_rename_img.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
